Interfaces -> Allow the serial device to be closed.
When running tests on the BV4626 device, it appears not to initialise correctly the
second time it is opened. This is a result of not closing it cleanly when finished,
so ensure this is possible.

It may make sense to do this automatically on object destruction but I did not want
to disrupt other devices that I cannot test